+++ Ranger/Extensions/String+Extensions.swift
@@ -19,6 +19,39 @@ extension String
         NSPasteboard.general.declareTypes([NSPasteboard.PasteboardType.string], owner: nil)
         NSPasteboard.general.setString(self, forType: NSPasteboard.PasteboardType.string)
     }
+    
+    func contains(each strings: [String]) -> Bool
+    {
+        strings.reduce(
+            true,
+            {
+                contains, eachString in
+                contains && self.contains(eachString)
+            }
+        )
+    }
+    
+    func contains(any strings: [String]) -> Bool
+    {
+        strings.reduce(
+            true,
+            {
+                contains, eachString in
+                contains && self.contains(eachString)
+            }
+        )
+    }
+    
+    func contained(in strings: [String]) -> Bool
+    {
+        strings.reduce(
+            false,
+            {
+                contains, eachString in
+                contains || self == eachString
+            }
+        )
+    }
 }
 
 extension Optional where Wrapped == String