+++ Ranger/UI/TournamentTableWindow/TournamentWindowController.swift
@@ -27,14 +27,23 @@ class TournamentWindowController: NSWindowController
         
         // Track table.
         let tourneyTableViewController = tournamentTableWindowController.contentViewController as! TournamentViewController
-            tourneyTableViewController.track(tableWindowInfo)
+            tourneyTableViewController.update(with: tableWindowInfo)
         
         // Inject reference.
         tournamentTableWindowController.tableWindowInfo = tableWindowInfo
         
         // Hide buttons.
-        if let window = tourneyTableViewController.view.window
+        if let window = tourneyTableViewController.view.window as? NSPanel
         {
+            // Styles.
+            window.styleMask = [.borderless, .titled, .closable, .nonactivatingPanel]
+            
+            // See-trough.
+            window.titlebarAppearsTransparent = true
+            window.backgroundColor = NSColor.clear
+            window.isOpaque = false
+         
+            // Hide controls.
             window.standardWindowButton(.miniaturizeButton)?.isHidden = true
             window.standardWindowButton(.zoomButton)?.isHidden = true
             window.standardWindowButton(.closeButton)?.isHidden = true
@@ -46,6 +55,15 @@ class TournamentWindowController: NSWindowController
     func update(with tableWindowInfo: TableWindowInfo)
     {
         let tourneyTableViewController = self.contentViewController as! TournamentViewController
-            tourneyTableViewController.update(with: tableWindowInfo)
+        
+        // Hide buttons.
+        if let window = tourneyTableViewController.view.window as? NSPanel
+        { window.order(NSWindow.OrderingMode.above, relativeTo: tableWindowInfo.number) }
+            
+        // window.makeKeyAndOrderFront(window)
+        // window.orderFrontRegardless()
+        
+        // Update content.
+        tourneyTableViewController.update(with: tableWindowInfo)
     }
 }